+++ glidertools/processing.py
@@ -149,6 +149,7 @@ def calc_oxygen(
 
     """
 
+    import numpy as np
     import seawater as sw
 
     from numpy import abs, array, c_, isnan, median, ones
@@ -202,7 +203,10 @@ def calc_oxygen(
         Y = o2raw[surf].copy()
         X = c_[ones(surf.sum()), o2sat[surf]]
         # removing outliers accodring to IQR
-        ll, ul = outlier_bounds_iqr(Y, multiplier=1.5)
+        # ll, ul = outlier_bounds_iqr(Y, multiplier=1.5)
+        iqr = outlier_bounds_iqr(Y, multiplier=1.5)
+        ll = np.nanmin(iqr)
+        ul = np.nanmax(iqr)
         m = (Y > ll) & (Y < ul)
         ratios = Y[m] / X[m, 1]
 
@@ -261,7 +265,7 @@ def calc_oxygen(
         o2aou = transfer_nc_attrs(
             getframe(),
             var,
-            o2mll,
+            o2aou,
             "o2aou",
             units="mL/L",
             comment="",
@@ -270,7 +274,7 @@ def calc_oxygen(
         o2pct = transfer_nc_attrs(
             getframe(),
             var,
-            o2mll,
+            o2pct,
             "o2pct",
             units="percent",
             comment="",